MQTT_Server/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
influxdb==5.3.1
msgpack==1.0.2
paho-mqtt==1.5.1
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6

celery~=5.1.2